RavenDB-26357 - Separate generators classes
