+++ test/SlowTests/Server/Documents/AI/AiAgent/RavenDB-26357.cs
@@ -40,7 +40,6 @@ public class RavenDB_26357 : RavenTestBase
         var result = await store.Maintenance.SendAsync(new GenerateCodeAiAgentsOperation(agentId, "c#"));
 
         var generatedCode = result.GeneratedCode;
-        generatedCode = generatedCode.Replace("/* your response type */", "string");
         generatedCode = generatedCode.Replace("using Raven.Client.Documents.Operations.AI.Agents;", """
 using Raven.Client.Documents.Operations.AI.Agents; 
 public class Program
